document_service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
python-docx
aiofiles
pydantic
chromadb
sentence-transformers
pytest
httpx
openai
pytesseract
Pillow